# api/swagger.yaml
openapi: 3.0.3
info:
  title: HW4 Orders & Payments API (Gateway)
  version: "1.0"
  description: |
    API для ДЗ-4.
    Gateway проксирует запросы в Orders Service и Payments Service.

servers:
  - url: http://localhost:8080

paths:
  /health:
    get:
      summary: Healthcheck
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
              example: ok

  /orders:
    post:
      summary: Create order
      description: |
        Создаёт заказ и асинхронно запускает оплату через Kafka.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
            examples:
              example:
                value:
                  user_id: "11111111-1111-1111-1111-111111111111"
                  amount: 600
                  description: "New Year gifts"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessCreateOrderResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      summary: List orders by user
      parameters:
        - in: query
          name: user_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessOrdersListResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{order_id}:
    get:
      summary: Get order by id
      parameters:
        - in: path
          name: order_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessOrderResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /accounts:
    post:
      summary: Create account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
            examples:
              example:
                value:
                  user_id: "11111111-1111-1111-1111-111111111111"
                  balance: 1000
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessGenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /accounts/topup:
    post:
      summary: Top up balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopUpRequest"
            examples:
              example:
                value:
                  user_id: "11111111-1111-1111-1111-111111111111"
                  amount: 500
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessGenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /accounts/{user_id}:
    get:
      summary: Get balance
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessBalanceResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"

    SuccessGenericResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: true
      example:
        data:
          status: ok

    CreateOrderRequest:
      type: object
      required: [user_id, amount]
      properties:
        user_id:
          type: string
          format: uuid
        amount:
          type: integer
          format: int64
          minimum: 1
        description:
          type: string

    CreateOrderResponse:
      type: object
      required: [order_id, status]
      properties:
        order_id:
          type: string
          format: uuid
        status:
          type: string
          description: Order status at creation time (NEW)

    SuccessCreateOrderResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/CreateOrderResponse"

    OrderResponse:
      type: object
      required: [order_id, amount, status, created_at]
      properties:
        order_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          nullable: true
        amount:
          type: integer
          format: int64
        status:
          type: string
          description: NEW | FINISHED | FAILED
        created_at:
          type: string
          format: date-time
        description:
          type: string
          nullable: true

    OrdersListResponse:
      type: object
      required: [orders]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/OrderResponse"

    SuccessOrdersListResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/OrdersListResponse"

    SuccessOrderResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/OrderResponse"

    CreateAccountRequest:
      type: object
      required: [user_id, balance]
      properties:
        user_id:
          type: string
          format: uuid
        balance:
          type: integer
          format: int64
          minimum: 0

    TopUpRequest:
      type: object
      required: [user_id, amount]
      properties:
        user_id:
          type: string
          format: uuid
        amount:
          type: integer
          format: int64
          minimum: 1

    BalanceResponse:
      type: object
      required: [user_id, balance]
      properties:
        user_id:
          type: string
          format: uuid
        balance:
          type: integer
          format: int64

    SuccessBalanceResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/BalanceResponse"